[distributed] skip use_ignore_output tests in c10d if not built with gloo (#33513)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33513

These tests require gloo so like the other tests, they should be
skipped if not building with gloo. Otherwise they crash on Mac if not built
with gloo currently.

verified that it does not crash anymore with this PR.
ghstack-source-id: 99303707

Test Plan: Built on Mac and verified that the tests do not fail.

Differential Revision: D19976908

fbshipit-source-id: 6a2a70c3eab83efd0e188e86cabe56de4a869f4d