[ci] try to fix rocm builds (#34600)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34600

They are failing with:
```
E: The method driver /usr/lib/apt/methods/https could not be found.
```

Trying the solution recommended in: https://unix.stackexchange.com/questions/263801/apt-get-fails-the-method-driver-usr-lib-apt-methods-https-could-not-be-found

The long-term solution is to move all this to be pre-installed in the
docker image.

Test Plan: Imported from OSS

Differential Revision: D20391153

Pulled By: suo

fbshipit-source-id: 959dff2ea9e77bb52739c0659e9d800cdbe4cb01