[workflows] Don't pipe clang_format.py output to a file (#35496)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35496

This commit modifies the clang-format workflow so that it prints
the output of `tools/clang_format.py` to stdout instead of piping
it to a file. This way, the issues encountered by the script
(e.g. which files are not formatted correctly) will be visible
in the CI window.

Testing:
CI

Test Plan: Imported from OSS

Differential Revision: D20678729

Pulled By: SplitInfinity

fbshipit-source-id: 8b437c2cf2779de0245c1b4301c57b4ee0dcad6d