[doc] Fix broken links in the TOC of CONTRIBUTING.md (#37131)
Summary:
Some links in the TOC of CONTRIBUTING.md is broken since GitHub removes the invalid characters (e.g., `+` in C++) in the anchor link, while the existing TOC uses `-` for replacement.

This PR uses `-` instead of `*` and `+` for the bullet lists to make it consistent with README.md.
https://github.com/pytorch/pytorch/blob/b496dfb7e36f5294564b367516b6c71c5f3b90ca/README.md#L11-L18
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37131

Differential Revision: D21231299

Pulled By: zou3519

fbshipit-source-id: 8e7bb61550827ce97378d3428542e43612bac8e1